+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project does NOT adhere to [Semantic Versioning](https://semver.org/spe
 
 ### Added
 
+* qts.py, a library for interacting with the QNAP QTS API
 * endoflifedate.py (built automatically; version string only changes when products are added or deleted)
 * [Published on PyPI](https://pypi.org/project/linuxfabrik-lib/), you may now use `pip install linuxfabrik-lib`